mgr/dashboard: isolate the logic from controller to host service
Fixes: https://tracker.ceph.com/issues/51252
