+++ src/pybind/mgr/dashboard/tests/test_host.py
@@ -6,7 +6,8 @@ from unittest import mock
 from orchestrator import HostSpec, InventoryHost
 
 from .. import mgr
-from ..controllers.host import Host, HostUi, get_device_osd_map, get_hosts, get_inventories
+from ..controllers.host import Host, HostUi
+from ..services.host import HostService
 from ..tools import NotificationQueue, TaskManager
 from . import ControllerTestCase  # pylint: disable=no-name-in-module
 
@@ -126,8 +127,198 @@ class HostControllerTest(ControllerTestCase):
             self.assertIn('status', self.json_body())
             self.assertIn('addr', self.json_body())
 
-    @mock.patch('dashboard.controllers.host.add_host')
-    def test_add_host(self, mock_add_host):
+    @mock.patch('dashboard.services.host.time')
+    def test_identify_device(self, mock_time):
+        url = '{}/host-0/identify_device'.format(self.URL_HOST)
+        with patch_orch(True) as fake_client:
+            payload = {
+                'device': '/dev/sdz',
+                'duration': '1'
+            }
+            self._task_post(url, payload)
+            self.assertStatus(200)
+            mock_time.sleep.assert_called()
+            calls = [
+                mock.call('host-0', '/dev/sdz', 'ident', True),
+                mock.call('host-0', '/dev/sdz', 'ident', False),
+            ]
+            fake_client.blink_device_light.assert_has_calls(calls)
+
+    @mock.patch('dashboard.services.host.HostService.get_inventories')
+    def test_inventory(self, mock_get_inventories):
+        inventory_url = '{}/host-0/inventory'.format(self.URL_HOST)
+        with patch_orch(True):
+            tests = [
+                {
+                    'url': inventory_url,
+                    'inventories': [{'a': 'b'}],
+                    'refresh': None,
+                    'resp': {'a': 'b'}
+                },
+                {
+                    'url': '{}?refresh=true'.format(inventory_url),
+                    'inventories': [{'a': 'b'}],
+                    'refresh': "true",
+                    'resp': {'a': 'b'}
+                },
+                {
+                    'url': inventory_url,
+                    'inventories': [],
+                    'refresh': None,
+                    'resp': {}
+                },
+            ]
+            for test in tests:
+                mock_get_inventories.reset_mock()
+                mock_get_inventories.return_value = test['inventories']
+                self._get(test['url'])
+                mock_get_inventories.assert_called_once_with(['host-0'], test['refresh'])
+                self.assertEqual(self.json_body(), test['resp'])
+                self.assertStatus(200)
+
+        # list without orchestrator service
+        with patch_orch(False):
+            self._get(inventory_url)
+            self.assertStatus(503)
+
+
+class HostServiceTest(unittest.TestCase):
+    def test_get_hosts(self):
+        mgr.list_servers.return_value = [{
+            'hostname': 'node1'
+        }, {
+            'hostname': 'localhost'
+        }]
+        orch_hosts = [
+            HostSpec('node1', labels=['foo', 'bar']),
+            HostSpec('node2', labels=['bar'])
+        ]
+
+        with patch_orch(True, hosts=orch_hosts):
+            hosts = HostService.get_hosts()
+            self.assertEqual(len(hosts), 3)
+            checks = {
+                'localhost': {
+                    'sources': {
+                        'ceph': True,
+                        'orchestrator': False
+                    },
+                    'labels': []
+                },
+                'node1': {
+                    'sources': {
+                        'ceph': True,
+                        'orchestrator': True
+                    },
+                    'labels': ['bar', 'foo']
+                },
+                'node2': {
+                    'sources': {
+                        'ceph': False,
+                        'orchestrator': True
+                    },
+                    'labels': ['bar']
+                }
+            }
+            for host in hosts:
+                hostname = host['hostname']
+                self.assertDictEqual(host['sources'], checks[hostname]['sources'])
+                self.assertListEqual(host['labels'], checks[hostname]['labels'])
+
+    @mock.patch('dashboard.services.host.mgr.get')
+    def test_get_device_osd_map(self, mgr_get):
+        mgr_get.side_effect = lambda key: {
+            'osd_metadata': {
+                '0': {
+                    'hostname': 'node0',
+                    'devices': 'nvme0n1,sdb',
+                },
+                '1': {
+                    'hostname': 'node0',
+                    'devices': 'nvme0n1,sdc',
+                },
+                '2': {
+                    'hostname': 'node1',
+                    'devices': 'sda',
+                },
+                '3': {
+                    'hostname': 'node2',
+                    'devices': '',
+                }
+            }
+        }[key]
+
+        device_osd_map = HostService.get_device_osd_map()
+        mgr.get.assert_called_with('osd_metadata')
+        # sort OSD IDs to make assertDictEqual work
+        for devices in device_osd_map.values():
+            for host in devices.keys():
+                devices[host] = sorted(devices[host])
+        self.assertDictEqual(device_osd_map, {
+            'node0': {
+                'nvme0n1': [0, 1],
+                'sdb': [0],
+                'sdc': [1],
+            },
+            'node1': {
+                'sda': [2]
+            }
+        })
+
+    @mock.patch('dashboard.services.host.str_to_bool')
+    @mock.patch('dashboard.services.host.HostService.get_device_osd_map')
+    def test_get_inventories(self, mock_get_device_osd_map, mock_str_to_bool):
+        mock_get_device_osd_map.return_value = {
+            'host-0': {
+                'nvme0n1': [1, 2],
+                'sdb': [1],
+                'sdc': [2]
+            },
+            'host-1': {
+                'sdb': [3]
+            }
+        }
+        inventory = [
+            {
+                'name': 'host-0',
+                'addr': '1.2.3.4',
+                'devices': [
+                    {'path': 'nvme0n1'},
+                    {'path': '/dev/sdb'},
+                    {'path': '/dev/sdc'},
+                ]
+            },
+            {
+                'name': 'host-1',
+                'addr': '1.2.3.5',
+                'devices': [
+                    {'path': '/dev/sda'},
+                    {'path': 'sdb'},
+                ]
+            }
+        ]
+
+        with patch_orch(True, inventory=inventory) as orch_client:
+            mock_str_to_bool.return_value = True
+
+            hosts = ['host-0', 'host-1']
+            inventories = HostService.get_inventories(hosts, 'true')
+            mock_str_to_bool.assert_called_with('true')
+            orch_client.inventory.list.assert_called_once_with(hosts=hosts, refresh=True)
+            self.assertEqual(len(inventories), 2)
+            host0 = inventories[0]
+            self.assertEqual(host0['name'], 'host-0')
+            self.assertEqual(host0['addr'], '1.2.3.4')
+            self.assertEqual(host0['devices'][0]['osd_ids'], [1, 2])
+            self.assertEqual(host0['devices'][1]['osd_ids'], [1])
+            self.assertEqual(host0['devices'][2]['osd_ids'], [2])
+            host1 = inventories[1]
+            self.assertEqual(host1['name'], 'host-1')
+            self.assertEqual(host1['addr'], '1.2.3.5')
+            self.assertEqual(host1['devices'][0]['osd_ids'], [])
+            self.assertEqual(host1['devices'][1]['osd_ids'], [3])
+
+    def test_add_host(self):
         with patch_orch(True):
             payload = {
                 'hostname': 'node0',
@@ -137,6 +328,7 @@ class HostControllerTest(ControllerTestCase):
             }
             self._post(self.URL_HOST, payload, version='0.1')
             self.assertStatus(201)
+            HostService.add_host('node0', '192.0.2.0', 'mon', 'maintenance')
             mock_add_host.assert_called()
 
     def test_set_labels(self):
@@ -178,6 +370,8 @@ class HostControllerTest(ControllerTestCase):
             self._put('{}/node1'.format(self.URL_HOST), {'maintenance': True, 'force': True},
                       version='0.1')
             self.assertStatus(200)
+            self.assertHeader('Content-Type',
+                              'application/vnd.ceph.api.v0.1+json')
 
             # exit maintenance mode
             self._put('{}/node0'.format(self.URL_HOST), {'maintenance': True}, version='0.1')
@@ -190,70 +384,6 @@ class HostControllerTest(ControllerTestCase):
             self._put('{}/node0'.format(self.URL_HOST), {'maintenance': True}, version='0.1')
             self.assertStatus(503)
 
-    @mock.patch('dashboard.controllers.host.time')
-    def test_identify_device(self, mock_time):
-        url = '{}/host-0/identify_device'.format(self.URL_HOST)
-        with patch_orch(True) as fake_client:
-            payload = {
-                'device': '/dev/sdz',
-                'duration': '1'
-            }
-            self._task_post(url, payload)
-            self.assertStatus(200)
-            mock_time.sleep.assert_called()
-            calls = [
-                mock.call('host-0', '/dev/sdz', 'ident', True),
-                mock.call('host-0', '/dev/sdz', 'ident', False),
-            ]
-            fake_client.blink_device_light.assert_has_calls(calls)
-
-    @mock.patch('dashboard.controllers.host.get_inventories')
-    def test_inventory(self, mock_get_inventories):
-        inventory_url = '{}/host-0/inventory'.format(self.URL_HOST)
-        with patch_orch(True):
-            tests = [
-                {
-                    'url': inventory_url,
-                    'inventories': [{'a': 'b'}],
-                    'refresh': None,
-                    'resp': {'a': 'b'}
-                },
-                {
-                    'url': '{}?refresh=true'.format(inventory_url),
-                    'inventories': [{'a': 'b'}],
-                    'refresh': "true",
-                    'resp': {'a': 'b'}
-                },
-                {
-                    'url': inventory_url,
-                    'inventories': [],
-                    'refresh': None,
-                    'resp': {}
-                },
-            ]
-            for test in tests:
-                mock_get_inventories.reset_mock()
-                mock_get_inventories.return_value = test['inventories']
-                self._get(test['url'])
-                mock_get_inventories.assert_called_once_with(['host-0'], test['refresh'])
-                self.assertEqual(self.json_body(), test['resp'])
-                self.assertStatus(200)
-
-        # list without orchestrator service
-        with patch_orch(False):
-            self._get(inventory_url)
-            self.assertStatus(503)
-
-
-class HostUiControllerTest(ControllerTestCase):
-    URL_HOST = '/ui-api/host'
-
-    @classmethod
-    def setup_server(cls):
-        # pylint: disable=protected-access
-        HostUi._cp_config['tools.authenticate.on'] = False
-        cls.setup_controllers([HostUi])
-
     def test_labels(self):
         orch_hosts = [
             HostSpec('node1', labels=['foo']),
@@ -261,172 +391,6 @@ class HostUiControllerTest(ControllerTestCase):
         ]
 
         with patch_orch(True, hosts=orch_hosts):
-            self._get('{}/labels'.format(self.URL_HOST))
-            self.assertStatus(200)
-            labels = self.json_body()
+            labels = HostService.get_labels()
             labels.sort()
             self.assertListEqual(labels, ['bar', 'foo'])
-
-    @mock.patch('dashboard.controllers.host.get_inventories')
-    def test_inventory(self, mock_get_inventories):
-        inventory_url = '{}/inventory'.format(self.URL_HOST)
-        with patch_orch(True):
-            tests = [
-                {
-                    'url': inventory_url,
-                    'refresh': None
-                },
-                {
-                    'url': '{}?refresh=true'.format(inventory_url),
-                    'refresh': "true"
-                },
-            ]
-            for test in tests:
-                mock_get_inventories.reset_mock()
-                mock_get_inventories.return_value = [{'a': 'b'}]
-                self._get(test['url'])
-                mock_get_inventories.assert_called_once_with(None, test['refresh'])
-                self.assertEqual(self.json_body(), [{'a': 'b'}])
-                self.assertStatus(200)
-
-        # list without orchestrator service
-        with patch_orch(False):
-            self._get(inventory_url)
-            self.assertStatus(503)
-
-
-class TestHosts(unittest.TestCase):
-    def test_get_hosts(self):
-        mgr.list_servers.return_value = [{
-            'hostname': 'node1'
-        }, {
-            'hostname': 'localhost'
-        }]
-        orch_hosts = [
-            HostSpec('node1', labels=['foo', 'bar']),
-            HostSpec('node2', labels=['bar'])
-        ]
-
-        with patch_orch(True, hosts=orch_hosts):
-            hosts = get_hosts()
-            self.assertEqual(len(hosts), 3)
-            checks = {
-                'localhost': {
-                    'sources': {
-                        'ceph': True,
-                        'orchestrator': False
-                    },
-                    'labels': []
-                },
-                'node1': {
-                    'sources': {
-                        'ceph': True,
-                        'orchestrator': True
-                    },
-                    'labels': ['bar', 'foo']
-                },
-                'node2': {
-                    'sources': {
-                        'ceph': False,
-                        'orchestrator': True
-                    },
-                    'labels': ['bar']
-                }
-            }
-            for host in hosts:
-                hostname = host['hostname']
-                self.assertDictEqual(host['sources'], checks[hostname]['sources'])
-                self.assertListEqual(host['labels'], checks[hostname]['labels'])
-
-    @mock.patch('dashboard.controllers.host.mgr.get')
-    def test_get_device_osd_map(self, mgr_get):
-        mgr_get.side_effect = lambda key: {
-            'osd_metadata': {
-                '0': {
-                    'hostname': 'node0',
-                    'devices': 'nvme0n1,sdb',
-                },
-                '1': {
-                    'hostname': 'node0',
-                    'devices': 'nvme0n1,sdc',
-                },
-                '2': {
-                    'hostname': 'node1',
-                    'devices': 'sda',
-                },
-                '3': {
-                    'hostname': 'node2',
-                    'devices': '',
-                }
-            }
-        }[key]
-
-        device_osd_map = get_device_osd_map()
-        mgr.get.assert_called_with('osd_metadata')
-        # sort OSD IDs to make assertDictEqual work
-        for devices in device_osd_map.values():
-            for host in devices.keys():
-                devices[host] = sorted(devices[host])
-        self.assertDictEqual(device_osd_map, {
-            'node0': {
-                'nvme0n1': [0, 1],
-                'sdb': [0],
-                'sdc': [1],
-            },
-            'node1': {
-                'sda': [2]
-            }
-        })
-
-    @mock.patch('dashboard.controllers.host.str_to_bool')
-    @mock.patch('dashboard.controllers.host.get_device_osd_map')
-    def test_get_inventories(self, mock_get_device_osd_map, mock_str_to_bool):
-        mock_get_device_osd_map.return_value = {
-            'host-0': {
-                'nvme0n1': [1, 2],
-                'sdb': [1],
-                'sdc': [2]
-            },
-            'host-1': {
-                'sdb': [3]
-            }
-        }
-        inventory = [
-            {
-                'name': 'host-0',
-                'addr': '1.2.3.4',
-                'devices': [
-                    {'path': 'nvme0n1'},
-                    {'path': '/dev/sdb'},
-                    {'path': '/dev/sdc'},
-                ]
-            },
-            {
-                'name': 'host-1',
-                'addr': '1.2.3.5',
-                'devices': [
-                    {'path': '/dev/sda'},
-                    {'path': 'sdb'},
-                ]
-            }
-        ]
-
-        with patch_orch(True, inventory=inventory) as orch_client:
-            mock_str_to_bool.return_value = True
-
-            hosts = ['host-0', 'host-1']
-            inventories = get_inventories(hosts, 'true')
-            mock_str_to_bool.assert_called_with('true')
-            orch_client.inventory.list.assert_called_once_with(hosts=hosts, refresh=True)
-            self.assertEqual(len(inventories), 2)
-            host0 = inventories[0]
-            self.assertEqual(host0['name'], 'host-0')
-            self.assertEqual(host0['addr'], '1.2.3.4')
-            self.assertEqual(host0['devices'][0]['osd_ids'], [1, 2])
-            self.assertEqual(host0['devices'][1]['osd_ids'], [1])
-            self.assertEqual(host0['devices'][2]['osd_ids'], [2])
-            host1 = inventories[1]
-            self.assertEqual(host1['name'], 'host-1')
-            self.assertEqual(host1['addr'], '1.2.3.5')
-            self.assertEqual(host1['devices'][0]['osd_ids'], [])
-            self.assertEqual(host1['devices'][1]['osd_ids'], [3])
